fix: copy font.ttf only if it does not already exist

+++ modules/core/fontconfig/default.nix
@@ -16,8 +16,10 @@ util.mkToggledModule [ "core" ] {
       in
       lib.mkIf flags.isNixOnDroid (
         lib.hm.dag.entryAfter [ "filesChanged" ] ''
-          run mkdir -p "$HOME/.termux"
-          run cp '${font}' "$HOME/.termux/font.ttf"
+          if ! test -e "$HOME/.termux/font.ttf"; then
+            run mkdir -p "$HOME/.termux"
+            run cp '${font}' "$HOME/.termux/font.ttf"
+          fi
         ''
       );
   };
